refactor API related code to keep it closer to each app
