+++ smbportal/profiles/api/fields.py
@@ -1,0 +1,62 @@
+#########################################################################
+#
+# Copyright 2019, GeoSolutions Sas.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+#
+#########################################################################
+
+"""Serializers for the smbportal REST API"""
+
+import logging
+
+from rest_framework.reverse import reverse
+from rest_framework import serializers
+
+logger = logging.getLogger(__name__)
+
+
+class SmbUserHyperlinkedIdentityField(serializers.HyperlinkedIdentityField):
+    """Custom serializer field to support showing a custom id for ``SmbUser``
+
+    This field allows using the keycloak UUID as the user's identity field in
+    the API
+
+    """
+
+    def get_url(self, obj, view_name, request, format):
+        return reverse(
+            view_name,
+            kwargs={
+                "uuid": obj.keycloak.UID
+            },
+            request=request,
+            format=format
+        )
+
+
+class SmbUserHyperlinkedRelatedField(serializers.HyperlinkedRelatedField):
+    """Custom serializer field to support showing a custom id for ``SmbUser``
+
+    This field allows using the keycloak UUID as the related attribute when
+    referencing user model from other models in the API
+
+    """
+
+    def get_url(self, obj, view_name, request, format):
+        return reverse(
+            view_name,
+            kwargs={
+                "uuid": obj.keycloak.UID
+            },
+            request=request,
+            format=format
+        )
+
+    def get_object(self, view_name, view_args, view_kwargs):
+        return self.get_queryset().get(keycloak__UID=view_kwargs.get("pk"))
+
+    def use_pk_only_optimization(self):
+        return False
