+++ smbportal/vehiclemonitor/api/serializers.py
@@ -1,0 +1,83 @@
+#########################################################################
+#
+# Copyright 2019, GeoSolutions Sas.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+#
+#########################################################################
+
+"""Serializers for the smbportal REST API"""
+
+import logging
+
+from rest_framework import serializers
+from rest_framework_gis.serializers import GeoFeatureModelSerializer
+
+import vehicles.models
+import vehiclemonitor.models
+
+logger = logging.getLogger(__name__)
+
+
+class BikeObservationSerializer(GeoFeatureModelSerializer):
+    url = serializers.HyperlinkedIdentityField(
+        view_name="api:bike-observations-detail",
+    )
+    bike = serializers.SlugRelatedField(
+        slug_field="short_uuid",
+        queryset=vehicles.models.Bike.objects.all()
+    )
+    bike_url = serializers.HyperlinkedRelatedField(
+        source="bike",
+        read_only=True,
+        view_name="api:bikes-detail",
+        lookup_field="short_uuid",
+    )
+
+    class Meta:
+        model = vehiclemonitor.models.BikeObservation
+        geo_field = "position"
+        fields = (
+            "url",
+            "id",
+            "bike",
+            "bike_url",
+            "reporter_id",
+            "reporter_type",
+            "reporter_name",
+            "address",
+            "created_at",
+            "observed_at",
+            "details",
+        )
+
+
+class MyBikeObservationSerializer(BikeObservationSerializer):
+    bike = serializers.SlugRelatedField(
+        slug_field="short_uuid",
+        queryset=vehicles.models.Bike.objects.all()
+    )
+    bike_url = serializers.HyperlinkedRelatedField(
+        source="bike",
+        read_only=True,
+        view_name="api:my-bikes-detail",
+        lookup_field="short_uuid",
+    )
+
+    class Meta:
+        model = vehiclemonitor.models.BikeObservation
+        geo_field = "position"
+        fields = (
+            "id",
+            "bike",
+            "bike_url",
+            "reporter_id",
+            "reporter_type",
+            "reporter_name",
+            "address",
+            "created_at",
+            "observed_at",
+            "details",
+        )
