+++ smbportal/vehicles/api/filters.py
@@ -1,6 +1,6 @@
 #########################################################################
 #
-# Copyright 2018, GeoSolutions Sas.
+# Copyright 2019, GeoSolutions Sas.
 # All rights reserved.
 #
 # This source code is licensed under the BSD-style license found in the
@@ -10,9 +10,7 @@
 
 from django_filters import rest_framework as django_filters
 
-import profiles.models
-import vehicles.models
-import vehiclemonitor.models
+from .. import models
 
 
 class BikeFilterSet(django_filters.FilterSet):
@@ -25,29 +23,10 @@ class BikeFilterSet(django_filters.FilterSet):
     )
 
     class Meta:
-        model = vehicles.models.Bike
+        model = models.Bike
         fields = [
             "short_uuid",
             "tag",
         ]
 
 
-class BikeObservationFilterSet(django_filters.FilterSet):
-    bike = django_filters.CharFilter(
-        field_name="bike__short_uuid",
-        lookup_expr="istartswith",
-    )
-
-    class Meta:
-        model = vehiclemonitor.models.BikeObservation
-        fields = [
-            "bike",
-        ]
-
-
-class SmbUserFilterSet(django_filters.FilterSet):
-    class Meta:
-        model = profiles.models.SmbUser
-        fields = [
-            "username"
-        ]
